requirements.txt
# openai-whisper 
ffmpeg-python
faster-whisper


# + needs to install: "brew install pkg-config"